Moved service-related logic into a separate module
