+++ src/python_pachyderm/mixin/__init__.py
@@ -5,9 +5,8 @@ exclusively in order to provide better code organization (because we have
 several mixins, rather than one giant `Client` class.)
 """
 
-from python_pachyderm.mixin.pfs import PFSMixin
-from python_pachyderm.mixin.pps import PPSMixin
-from python_pachyderm.mixin.transaction import TransactionMixin
-from python_pachyderm.mixin.version import VersionMixin
-from python_pachyderm.mixin.admin import AdminMixin
-from python_pachyderm.mixin.util import Service, GRPC_MODULES, PROTO_MODULES
+from .pfs import PFSMixin
+from .pps import PPSMixin
+from .transaction import TransactionMixin
+from .version import VersionMixin
+from .admin import AdminMixin
