+++ src/python_pachyderm/mixin/version.py
@@ -1,4 +1,4 @@
-from .util import Service
+from python_pachyderm.service import Service
 from python_pachyderm.proto.version.versionpb import version_pb2_grpc as version_grpc
